# ☄️ Improve performance by removing all spread/rest usages