| Q                 | A
| ----------------- | ---
| Bug fix?          | no
| Breaking change?  | no
| New feature?      | no
| Deprecations?     | no
| Spec compliancy?  | no
| Tests added/pass? | yes
| Fixed tickets     | 
| License           | MIT

This removes all spread/rest usages from flow and estree plugin which improves the performance on node 6.10.3 significantly.

Before:
```
name                            run  
./fixtures/angular.js           352ms
./fixtures/backbone.js          33ms 
./fixtures/ember.debug.js       933ms
./fixtures/jquery.js            183ms
./fixtures/react-with-addons.js 377ms
```

After:

```
name                            run  
./fixtures/angular.js           273ms
./fixtures/backbone.js          24ms 
./fixtures/ember.debug.js       707ms
./fixtures/jquery.js            141ms
./fixtures/react-with-addons.js 298ms
```

Test repo: https://github.com/danez/babylon_performance

Also had to fix some flow errors.